+++ R/packagefinder.r
@@ -31,7 +31,7 @@ NULL
 ###   PACKAGE PACKAGEFINDER
 ###
 ###   Author and maintainer: Joachim Zuckarelli (joachim@zuckarelli.de)
-###   Version 0.1.4
+###   Version 0.1.5
 ###
 
 
@@ -191,12 +191,15 @@ buildIndex <- function(filename="", download.stats = FALSE) {
 #' @param limit.results The maximum number of matches presented in the search results; choose a negative number to display all results
 #' @param silent Indicates whether details of the user's search query are repeated in the console.
 #' @param index Either a path (or URL) to a search index, or a search index that is already loaded. If no index is provided, \code{findPackage()} creates an ad hoc search index.
-
+#' @param advanced.ranking Indicates if the ranking of search results shall be based on weights taking into account the inverse frequencies of the different search terms across all packages and the length of the matches relative to the texts they were found in. Usually, using advanced ranking (\code{advanced.ranking = TRUE}, default) gives more relevant results, especially in \code{"or"} mode when the search terms differ strongly in their frequency of occurrence across packages.
+#'
 #' @return No return value.
 #'
 #' @details The \code{GO} column in the search results is an index number that can be used to address the found package easily with the \code{go()} function.
 #' The \code{DOWNL_TOTAL} column in the search results gives the overall number of downloads of the respective package since its submission to CRAN. The number is based on the figures for the RStudio CRAN mirror server.
 #' This field is only provided if the search index contains download figures. Ad hoc indices (when \code{index = NULL}) never include download statistics. Please refer to \code{\link{buildIndex}()} for more details.
+#'
+#' \code{\link{fp}()} is a shorter alias for \code{\link{findPackage}()}.
 #'
 #' @author Joachim Zuckarelli \email{joachim@@zuckarelli.de}
 #'
@@ -212,7 +215,7 @@ buildIndex <- function(filename="", download.stats = FALSE) {
 #'    always.sensitive="GLM", index=searchindex)
 #' }
 #' @export
-findPackage<-function(keywords, mode = "or", case.sensitive = FALSE, always.sensitive = NULL, weights = c(1,2,1,2), display = "viewer", results.longdesc = FALSE, limit.results = 15, silent = FALSE, index = NULL) {
+findPackage<-function(keywords, mode = "or", case.sensitive = FALSE, always.sensitive = NULL, weights = c(2,2,1,2), display = "viewer", results.longdesc = FALSE, limit.results = 15, silent = FALSE, index = NULL, advanced.ranking = TRUE) {
   if(!silent) {
     if(mode=="and") {
       mode.param <- "and"
@@ -238,36 +241,63 @@ findPackage<-function(keywords, mode = "or", case.sensitive = FALSE, always.sens
   searchindex <- makeIndexAvailable(index)
   if(class(searchindex) == "list"){
     score<-c()
+    m = matrix(nrow=length(searchindex$index$NAME), ncol=length(keywords)+length(always.sensitive))
+    m1.lengthratio <- 1
+    m2.lengthratio <- 1
+    m3.lengthratio <- 1
+
+    num.keywords = length(keywords)
+    if(!is.null(always.sensitive)) keywords = c(keywords, always.sensitive)
+
     for(i in 1:length(searchindex$index$NAME)) {
-      found<-c()
-      m1<-0
-      m2<-0
-      m3<-0
+
       for(f in 1:length(keywords)) {
-        dm1 <- stringr::str_count(searchindex$index$NAME[i], stringr::fixed(keywords[f], ignore_case=!case.sensitive))
-        dm2 <- stringr::str_count(searchindex$index$DESC_SHORT[i], stringr::fixed(keywords[f], ignore_case=!case.sensitive))
-        dm3 <- stringr::str_count(searchindex$index$DESC_LONG[i], stringr::fixed(keywords[f], ignore_case=!case.sensitive))
-        m1 <- m1 + dm1 * weights[1]
-        m2 <- m2 + dm2 * weights[2]
-        m3 <- m3 + dm3 * weights[3]
-        found[f] <- !(dm1 + dm2 + dm3 == 0)
-      }
+        m1<-0
+        m2<-0
+        m3<-0
 
-      if(!is.null(always.sensitive)) {
-        for(f in 1:length(always.sensitive)) {
-          dm1 <- stringr::str_count(searchindex$index$NAME[i], stringr::fixed(always.sensitive[f], ignore_case=FALSE))
-          dm2 <- stringr::str_count(searchindex$index$DESC_SHORT[i], stringr::fixed(always.sensitive[f], ignore_case=FALSE))
-          dm3 <- stringr::str_count(searchindex$index$DESC_LONG[i], stringr::fixed(always.sensitive[f], ignore_case=FALSE))
-          m1 <- m1 + dm1 * weights[1]
-          m2 <- m2 + dm2 * weights[2]
-          m3 <- m3 + dm3 * weights[3]
-          found[length(keywords)+f] <- !(dm1 + dm2 + dm3 == 0)
+        if(f <= num.keywords) {
+          cs <- !case.sensitive
+        } else {
+          cs <- FALSE
         }
+        dm1 <- stringr::str_count(searchindex$index$NAME[i], stringr::fixed(keywords[f], ignore_case=cs))
+        dm2 <- stringr::str_count(searchindex$index$DESC_SHORT[i], stringr::fixed(keywords[f], ignore_case=cs))
+        dm3 <- stringr::str_count(searchindex$index$DESC_LONG[i], stringr::fixed(keywords[f], ignore_case=cs))
+        if(advanced.ranking) {
+          m1.lengthratio <- dm1 * nchar(keywords[f]) / nchar(searchindex$index$NAME[i])
+          m2.lengthratio <- dm2 * nchar(keywords[f]) / nchar(searchindex$index$DESC_SHORT[i])
+          m3.lengthratio <- dm3 * nchar(keywords[f]) / nchar(searchindex$index$DESC_LONG[i])
+        }
+        m1 <- dm1 * weights[1] * m1.lengthratio
+        m2 <- dm2 * weights[2] * m2.lengthratio
+        m3 <- dm3 * weights[3] * m3.lengthratio
+        m[i,f] <- m1 + m2 + m3
       }
-
-      if(mode == "or") score[i] <- (m1 + m2 + m3) * (weights[4] - (weights[4]-1) * (length(found[found==FALSE])!=0))
-      else score[i] <- (m1 + m2 + m3) * weights[4] * (length(found[found==FALSE])==0)
     }
+
+    m <- m / max(m, na.rm=TRUE)
+
+    inverse.keyword.weight <- c()
+    if(!advanced.ranking) {
+      inverse.keyword.weight <- rep(1, length(keywords))
+    } else {
+      for(f in 1:length(keywords)) {
+        inverse.keyword.weight[f] <- sum(m[,f]>0, na.rm=TRUE)
+      }
+      inverse.keyword.weight <- 1 / (inverse.keyword.weight / max(inverse.keyword.weight, na.rm=TRUE))
+    }
+
+    if(mode == "or") {
+      for(i in 1:length(searchindex$index$NAME)) {
+        score[i] <- (sum(m[i,] * inverse.keyword.weight, na.rm=TRUE) * (1 + (weights[4] - 1) * (sum(m[i,]==0, na.rm=TRUE)==0)))
+      }
+    } else {
+      for(i in 1:length(searchindex$index$NAME)) {
+        score[i] <- (sum(m[i,] * inverse.keyword.weight, na.rm=TRUE) * (sum(m[i,]==0, na.rm=TRUE)==0))
+      }
+    }
+
     score <- score / max(score, na.rm=TRUE) * 100
     searchindex$index$SCORE <- as.numeric(score)
 
@@ -281,7 +311,7 @@ findPackage<-function(keywords, mode = "or", case.sensitive = FALSE, always.sens
       } else {
         # res.cols <- c("SCORE", "NAME", "DESC_SHORT", "DOWNL_TOTAL", "GO")
         res.cols <-c(26,1,3,7)
-       text.align.formattable <- c("r", "l", "l", "l", "l")
+        text.align.formattable <- c("r", "l", "l", "l", "l")
         text.align.pandoc <- c("right", "left", "left", "left", "left")
       }
 
@@ -316,6 +346,24 @@ findPackage<-function(keywords, mode = "or", case.sensitive = FALSE, always.sens
     stop("Search index is not available. Create a search index with buidIndex().")
   }
 }
+
+
+
+
+#' @title Searching for packages on CRAN
+#' @description Shorter alias for function \code{\link{findPackage}()}.
+#'
+#' @param ... Arguments as in \code{\link{findPackage}()}.
+#'
+#' @author Joachim Zuckarelli \email{joachim@@zuckarelli.de}
+#'
+#' @examples
+#' \donttest{
+#' fp(c("meta", "regression"))
+#' }
+#' @export
+fp <- function(...) { findPackage(...) }
+
 
 
 #' @title Searching for packages on CRAN
@@ -517,7 +565,9 @@ lastResults <- function(display = "viewer") {
 #'
 #' @author Joachim Zuckarelli \email{joachim@@zuckarelli.de}
 #' @examples
+#' \donttest{
 #' whatsNew(last.days = 3)
+#'}
 #'
 #' @export
 whatsNew <- function(last.days=0, brief = TRUE, index = NULL) {
@@ -544,7 +594,7 @@ whatsNew <- function(last.days=0, brief = TRUE, index = NULL) {
           num <- num + 1
         }
       }
-    invisible(num)
+      invisible(num)
     }
     else {
       stop("Argument last.days must be equal to or larger than zero.")